@@ -1,10 +1,12 @@
 <script>
-  export let slug = "";
-  export let image_scan = "";
-  export let title = "";
-  export let user_id = "";
-  export let categories = [];  
-  export let viewtransition = false;
+  let {
+    slug = "",
+    image_scan = "",
+    title = "",
+    user_id = "",
+    categories = [],
+    viewtransition = false
+  } = $props();
 
 </script>
 
