@@ -1,6 +1,6 @@
 <script>
-  export let data;
   import { Search, SearchResults } from "$lib/index.js";
+  let { data } = $props();
 </script>
 
 <main>
